refactor(Bundles): split CellsBundleController into a folder [YTFRONT-5892]
Separate the presentational column/render helpers (utils.tsx), the
useColumns hook, and the main component so the sorting feature can be
layered on cleanly. No behavior change.
