+++ packages/ui/src/ui/pages/tablet_cell_bundles/cells/CellsBundleController/index.ts
@@ -1,0 +1,1 @@
+export {CellsBundleController} from './CellsBundleController';
