+++ packages/ui/src/ui/pages/tablet_cell_bundles/cells/CellsBundleController/useColumns.ts
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import {COLUMNS} from './constants';
+
+export const useColumns = (hideColumns?: Array<string>) => {
+    const res = React.useMemo(() => {
+        if (!hideColumns?.length) {
+            return COLUMNS;
+        }
+
+        const toHide = new Set(hideColumns);
+        return COLUMNS.filter((item) => !toHide.has(item.name));
+    }, [hideColumns]);
+    return res;
+};
